backend/pyproject.toml
[tool.poetry]
name = "ragnroll"
version = "0.1.0"
description = ""
authors = ["Izhar Firdaus <kagesenshi.87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
neo4j = "^5.15.0"
uvicorn = "^0.26.0"
pydantic-settings = "^2.1.0"
langchain = "^0.1.2"
poetry-dotenv-plugin = "^0.2.0"
langchain-openai = "^0.0.5"
python-magic = "^0.4.27"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ragnroll = 'ragnroll.main:run'
